Substituting ChannelRepository
