+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/message/internal/RealmChannelRepository.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.ui.sample.realm.repository
+package io.getstream.chat.android.offline.repository.domain.message.internal
 
 import android.util.Log
 import io.getstream.chat.android.client.models.Channel
@@ -6,9 +6,9 @@ import io.getstream.chat.android.client.models.Member
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.persistance.repository.ChannelRepository
 import io.getstream.chat.android.client.utils.SyncStatus
-import io.getstream.chat.ui.sample.realm.entity.ChannelEntityRealm
-import io.getstream.chat.ui.sample.realm.entity.toDomain
-import io.getstream.chat.ui.sample.realm.entity.toRealm
+import io.getstream.chat.android.offline.repository.domain.channel.internal.ChannelEntityRealm
+import io.getstream.chat.android.offline.repository.domain.channel.internal.toDomain
+import io.getstream.chat.android.offline.repository.domain.channel.internal.toRealm
 import io.realm.kotlin.Realm
 import io.realm.kotlin.UpdatePolicy
 import io.realm.kotlin.ext.query
@@ -17,7 +17,7 @@ import java.util.*
 private const val LOG_TAG: String = "RealmChannelRepository"
 private const val NO_LIMIT: Int = -1
 
-internal class RealmChannelRepository(private val realm: Realm) : ChannelRepository {
+public class RealmChannelRepository(private val realm: Realm) : ChannelRepository {
 
   override suspend fun clear() {
     realm.writeBlocking {
